+++ source/Nuke.Azure.KeyVault/IconClasses.cs
@@ -1,0 +1,15 @@
+﻿// Copyright Sebastian Karasek, Matthias Koch 2018.
+// Distributed under the MIT License.
+// https://github.com/nuke-build/azure-keyvault/blob/master/LICENSE
+
+using System;
+using System.Linq;
+using Nuke.Azure.KeyVault;
+using Nuke.Common.Execution;
+
+[assembly: IconClass(typeof(KeyVaultTasks), "safe")]
+[assembly: IconClass(typeof(KeyVaultCertificateAttribute), "safe")]
+[assembly: IconClass(typeof(KeyVaultKeyAttribute), "safe")]
+[assembly: IconClass(typeof(KeyVaultAttribute), "safe")]
+[assembly: IconClass(typeof(KeyVaultSettingsAttribute), "safe")]
+[assembly: IconClass(typeof(KeyVaultSecretAttribute), "safe")]